+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `product-summary-quantity` not working.
+- Remove changes of `1.3.1`.
 
 ## [1.3.1] - 2020-04-03
 ### Fixed